Module prusti_interface::environment::borrowck::facts

Structs

BorrowckFacts

Interner

Point

A program point used in the borrow checker analysis.

Enums

PointType

The type of the point. Either the start of a statement or in the middle of it.

Type Definitions

AllInputFacts

AllOutputFacts

Loan

Path

PointIndex

Region

Variable