Module prusti_interface::environment::mir_analyses::initialization

This module provides the definitely initialized analysis for MIR.

Definitely initialized:

The working set S is the set of paths whose leaves are definitely initialized. For example, if we have x.f in S, then we know that x.f.g and x.f.h are definitely initialized. The invariant of this set is that we never have a node and any of its descendents in the set at the same time. For example, having x.f and x.f.g in S at the same time is illegal.

Structs

AnalysisResult

Functions

compute_definitely_initialized

Type Definitions

DefinitelyInitializedAnalysisResult

The result of the definitely initialized analysis.