Struct prusti_interface::environment::mir_analyses::initialization::AnalysisResult

pub struct AnalysisResult<T> {
    pub before_block: FxHashMap<BasicBlock, T>,
    pub after_statement: FxHashMap<Location, T>,
}

Fields
before_block: FxHashMap<BasicBlock, T>

The state before the basic block.

after_statement: FxHashMap<Location, T>

The state after the statement.

Implementations

impl<T> AnalysisResult<T>

pub fn new() -> Self

pub fn get_before_block(&self, bb: BasicBlock) -> &T

Get the initialization set before the first statement of the basic block.

pub fn get_after_statement(&self, location: Location) -> &T

Get the initialization set after the statement. If location.statement_index is equal to the number of statements, returns the initialization set after the terminator.