Type Alias prusti_interface::environment::mir_analyses::initialization::DefinitelyInitializedAnalysisResult

pub type DefinitelyInitializedAnalysisResult<'tcx> = AnalysisResult<PlaceSet<'tcx>>;

The result of the definitely initialized analysis.

Aliased Type

struct DefinitelyInitializedAnalysisResult<'tcx> {
    pub before_block: HashMap<BasicBlock, PlaceSet<'tcx>, BuildHasherDefault<FxHasher>>,
    pub after_statement: HashMap<Location, PlaceSet<'tcx>, BuildHasherDefault<FxHasher>>,
}

Fields
before_block: HashMap<BasicBlock, PlaceSet<'tcx>, BuildHasherDefault<FxHasher>>

The state before the basic block.

after_statement: HashMap<Location, PlaceSet<'tcx>, BuildHasherDefault<FxHasher>>

The state after the statement.

Implementations

impl<T> AnalysisResult<T>

pub fn new() -> Self

pub fn get_before_block(&self, bb: BasicBlock) -> &T

Get the initialization set before the first statement of the basic block.

pub fn get_after_statement(&self, location: Location) -> &T

Get the initialization set after the statement. If location.statement_index is equal to the number of statements, returns the initialization set after the terminator.