Struct prusti_interface::environment::mir_body::borrowck::lifetimes::Lifetimes

pub struct Lifetimes { /* private fields */ }

Implementations

impl Lifetimes

pub fn new(input_facts: AllInputFacts, location_table: LocationTable) -> Self

pub fn get_loan_live_at_start(&self, location: Location) -> BTreeSet<String>

pub fn get_loan_live_at_mid(&self, location: Location) -> BTreeSet<String>

pub fn get_origin_contains_loan_at_start(
&self,
location: Location
) -> BTreeMap<String, BTreeSet<String>>

pub fn get_origin_contains_loan_at_mid(
&self,
location: Location
) -> BTreeMap<String, BTreeSet<String>>

pub fn get_origin_contains_loan_at_location(
&self,
location: RichLocation
) -> BTreeMap<String, BTreeSet<String>>

pub fn lifetime_count(&self) -> usize

pub fn get_opaque_lifetimes_with_inclusions_names(
&self
) -> BTreeMap<String, BTreeSet<String>>

pub fn get_opaque_lifetimes_with_names(&self) -> Vec<String>

pub fn get_subset_base_at_start(
&self,
location: Location
) -> Vec<(Region, Region)>

pub fn get_subset_base_at_mid(&self, location: Location) -> Vec<(Region, Region)>

pub fn get_lifetimes_dead_on_edge(
&self,
from: RichLocation,
to: RichLocation
) -> Vec<Region>

pub fn construct_replacement_map(
&self,
location: Location,
base_lifetimes: Vec<LifetimeConst>
) -> BTreeMap<LifetimeConst, LifetimeConst>

When the Rust compiler generates Polonius facts, it creates many temporary lifetimes. Sometimes (for example, when encoding the constructing assignment of the aggregate with a reference-typed field) we need to unify the lifetimes. This method constructs a map from lifetimes known to Polonius to base_lifetimes indicating with which base_lifetimes the lifetimes could be replaced.