Struct prusti_interface::environment::mir_sets::place_set::PlaceSet

pub struct PlaceSet<'tcx> { /* private fields */ }

A set of MIR places.

Invariant: we never have a place and any of its descendants in the set at the same time. For example, having x.f and x.f.g in the set at the same time is illegal.

Implementations

impl<'tcx> PlaceSet<'tcx>

pub fn new() -> Self

pub fn contains(&self, place: Place<'tcx>) -> bool

pub fn contains_prefix_of(&self, place: Place<'tcx>) -> bool

pub fn check_invariant(&self)

pub fn iter<'a>(&'a self) -> impl Iterator<Item = &'a Place<'tcx>>

pub fn into_iter(self) -> impl Iterator<Item = Place<'tcx>>

pub fn insert(&mut self, place: Place<'tcx>, mir: &Body<'tcx>, tcx: TyCtxt<'tcx>)

Insert place.

pub fn remove(&mut self, place: Place<'tcx>, mir: &Body<'tcx>, tcx: TyCtxt<'tcx>)

Remove place.

pub fn merge(
place_set1: &PlaceSet<'tcx>,
place_set2: &PlaceSet<'tcx>
) -> PlaceSet<'tcx>

Compute the intersection of the two place sets.

pub fn deduplicate(&mut self)

This function fixes the invariant.

pub fn union(
place_set1: &PlaceSet<'tcx>,
place_set2: &PlaceSet<'tcx>
) -> PlaceSet<'tcx>

Compute the union of the two place sets.

Note that this function may break the invariant that we never have a place and its descendants in the set.