Struct prusti_interface::environment::polonius_info::AdditionalFacts

pub struct AdditionalFacts {
    pub loans: Vec<Loan>,
    pub reborrows: Vec<(Loan, Loan)>,
    pub reborrows_direct: Vec<(Loan, Loan)>,
    pub zombie_requires: FxHashMap<PointIndex, BTreeMap<Region, BTreeSet<Loan>>>,
    pub zombie_borrow_live_at: FxHashMap<PointIndex, Vec<Loan>>,
    pub borrow_become_zombie_at: FxHashMap<PointIndex, Vec<Loan>>,
}

Additional facts derived from the borrow checker facts.

Fields
loans: Vec<Loan>

A list of loans sorted by id.

reborrows: Vec<(Loan, Loan)>

The reborrows facts are needed for removing “fake” loans: at a specific program point there are often more than one loan active, but we are interested in only one of them, which is the original one. Therefore, we find all loans that are reborrows of the original loan and remove them. Reborrowing is defined as follows:

reborrows(Loan, Loan);
reborrows(L1, L2) :-
    loan_issued_at(R, L1, P),
    origin_contains_loan_at(R, P, L2).
reborrows(L1, L3) :-
    reborrows(L1, L2),
    reborrows(L2, L3).

reborrows_direct: Vec<(Loan, Loan)>

Non-transitive reborrows.

zombie_requires: FxHashMap<PointIndex, BTreeMap<Region, BTreeSet<Loan>>>

The zombie_requires facts are requires facts for the loans that were loan_killed_at.

zombie_requires(Region, Loan, Point);
zombie_requires(R, L, Q) :-
    requires(R, L, P),
    loan_killed_at(L, P),
    cfg_edge(P, Q),
    origin_live_on_entry(R, Q).
zombie_requires(R2, L, P) :-
    zombie_requires(R1, L, P),
    subset(R1, R2, P).
zombie_requires(R, L, Q) :-
    zombie_requires(R, L, P),
    cfg_edge(P, Q),
    origin_live_on_entry(R, Q).

zombie_borrow_live_at: FxHashMap<PointIndex, Vec<Loan>>

The zombie_borrow_live_at facts are loan_live_at facts for the loans that were loan_killed_at.

zombie_borrow_live_at(L, P) :-
    zombie_requires(R, L, P),
    origin_live_on_entry(R, P).

borrow_become_zombie_at: FxHashMap<PointIndex, Vec<Loan>>

Which loans were loan_killed_at (become zombies) at a given point.

Implementations

impl AdditionalFacts

pub fn new( all_facts: &AllInputFacts, output: &AllOutputFacts, incompatible_loans: &[Vec<Loan>] ) -> AdditionalFacts

Derive additional facts from the borrow checker facts.