Struct prusti_interface::environment::EnvQuery

pub struct EnvQuery<'tcx> { /* private fields */ }

Implementations

impl<'tcx> EnvQuery<'tcx>

pub fn new(tcx: TyCtxt<'tcx>) -> Self

pub fn hir(self) -> Map<'tcx>

Returns the hir context

pub fn codemap(self) -> &'tcx SourceMap

Returns the CodeMap

pub fn as_local_def_id(self, def_id: HirId) -> LocalDefId

Convert HirId to LocalDefId (see local_def_id in hir Map)

pub fn as_hir_id(self, def_id: LocalDefId) -> HirId

Convert LocalDefId to HirId (see local_def_id_to_hir_id in hir Map)

pub fn get_local_attributes(
self,
def_id: impl IntoParamTcx<'tcx, HirId>
) -> &'tcx [Attribute]

pub fn get_attributes(
self,
def_id: impl IntoParam<ProcedureDefId>
) -> &'tcx [Attribute]

pub fn has_prusti_attribute(
self,
def_id: impl IntoParam<ProcedureDefId>,
name: &str
) -> bool

Find whether the procedure has a particular prusti::<name> attribute.

pub fn get_def_span(self, def_id: impl IntoParam<DefId>) -> Span

Get the span of the given definition.

pub fn has_body(self, def_id: impl IntoParam<DefId>) -> bool

Returns true iff def_id has an MIR body which we may want to access

pub fn get_assoc_item(
self,
trait_id: impl IntoParam<DefId>,
item_id: impl IntoParam<DefId>
) -> Option<AssocItem>

Get an associated item by name.

pub fn get_trait_of_item(
self,
def_id: impl IntoParam<ProcedureDefId>
) -> Option<DefId>

If the given DefId describes an item belonging to a trait, returns the DefId of the trait that the trait item belongs to; otherwise, returns None.

pub fn is_trait_method_impl(self, def_id: impl IntoParam<ProcedureDefId>) -> bool

Returns true iff def_id is an implementation of a trait method

pub fn is_unsafe_function(self, def_id: impl IntoParam<ProcedureDefId>) -> bool

Returns true iff def_id is an unsafe function.

pub fn get_fn_sig(
self,
def_id: impl IntoParam<ProcedureDefId>,
substs: SubstsRef<'tcx>
) -> PolyFnSig<'tcx>

Computes the signature of the function with subst applied.

pub fn get_fn_sig_resolved(
self,
def_id: impl IntoParam<ProcedureDefId>,
substs: SubstsRef<'tcx>,
caller_def_id: impl IntoParam<ProcedureDefId>
) -> PolyFnSig<'tcx>

Computes the signature of the function with subst applied and associated types resolved.

pub fn is_closure(self, def_id: impl IntoParam<DefId>) -> bool

Returns true iff def_id is a closure.

pub fn find_trait_method_substs(
self,
impl_method_def_id: impl IntoParam<ProcedureDefId>,
impl_method_substs: SubstsRef<'tcx>
) -> Option<(ProcedureDefId, SubstsRef<'tcx>)>

If the given impl_method_def_id is an implementation of a trait method, return the DefId of that trait method as well as an adapted version of the callsite impl_method_substs substitutions.

pub fn find_impl_of_trait_method_call(
self,
proc_def_id: impl IntoParam<ProcedureDefId>,
substs: SubstsRef<'tcx>
) -> Option<ProcedureDefId>

Given some procedure proc_def_id which is called, this method returns the actual method which will be executed when proc_def_id is defined on a trait. Returns None if this method can not be found or the provided proc_def_id is no trait item.

pub fn resolve_method_call(
self,
caller_def_id: impl IntoParam<ProcedureDefId>,
called_def_id: impl IntoParam<ProcedureDefId>,
call_substs: SubstsRef<'tcx>
) -> (ProcedureDefId, SubstsRef<'tcx>)

Given a call to called_def_id from within caller_def_id, returns the DefId that will actually be called if known (i.e. if a trait method call actually resolves to a concrete implementation), as well as the correct substitutions for that call. If a method is not resolved, returns the original called_def_id and call_substs.

pub fn type_is_copy(
self,
ty: Binder<'tcx, Ty<'tcx>>,
param_env: impl IntoParamTcx<'tcx, ParamEnv<'tcx>>
) -> bool

Checks whether ty is copy. The type is wrapped into a Binder to handle regions correctly. The param_env should be passed as a ProcedureDefId which is then used to calculate the param env; i.e. the set of where-clauses that are in scope at this particular point.

pub fn type_implements_trait(
self,
trait_def_id: DefId,
ty: Ty<'tcx>,
param_env: impl IntoParamTcx<'tcx, ParamEnv<'tcx>>
) -> bool

Checks whether the given type implements the trait with the given DefId. The param_env should be passed as a ProcedureDefId which is then used to calculate the param env; i.e. the set of where-clauses that are in scope at this particular point.

pub fn type_implements_trait_with_trait_substs(
self,
trait_def_id: DefId,
trait_substs: impl IntoIterator<Item = impl Into<GenericArg<'tcx>>>,
param_env: impl IntoParamTcx<'tcx, ParamEnv<'tcx>>
) -> bool

Checks whether the given type implements the trait with the given DefId. Accounts for generic params on the trait by the given trait_substs. The param_env should be passed as a ProcedureDefId which is then used to calculate the param env; i.e. the set of where-clauses that are in scope at this particular point.

pub fn identity_substs(
self,
def_id: impl IntoParam<ProcedureDefId>
) -> SubstsRef<'tcx>

Return the default substitutions for a particular item, i.e. where each generic maps to itself.

pub fn evaluate_predicate(
self,
predicate: Predicate<'tcx>,
param_env: impl IntoParamTcx<'tcx, ParamEnv<'tcx>>
) -> bool

Evaluates the provided ty::Predicate. Returns true if the predicate is fulfilled. Returns false if the predicate is not fulfilled or it could not be evaluated. The param_env should be passed as a ProcedureDefId which is then used to calculate the param env; i.e. the set of where-clauses that are in scope at this particular point.

pub fn resolve_assoc_types<T: TypeFoldable<'tcx> + Debug + Copy>(
self,
normalizable: T,
param_env: impl IntoParamTcx<'tcx, ParamEnv<'tcx>>
) -> T

Normalizes associated types in foldable types, i.e. this resolves projection types ([ty::TyKind::Projection]s) Important: Regions while be erased during this process The param_env should be passed as a ProcedureDefId which is then used to calculate the param env; i.e. the set of where-clauses that are in scope at this particular point.