Struct prusti_interface::environment::Environment

pub struct Environment<'tcx> { /* fields omitted */ }

Facade to the Rust compiler.

Implementations

impl<'tcx> Environment<'tcx>

pub fn new(tcx: TyCtxt<'tcx>) -> Self

Builds an environment given a compiler state.

pub fn source_path(&self) -> PathBuf

Returns the path of the source that is being compiled

pub fn source_file_name(&self) -> String

Returns the file name of the source that is being compiled

pub fn crate_name(&self) -> String

Returns the name of the crate that is being compiled

pub fn tcx(&self) -> TyCtxt<'tcx>

Returns the typing context

pub fn hir_id_to_type(&self, hir_id: HirId) -> Ty<'tcx>

Returns the type of a HirId

pub fn codemap(&self) -> &'tcx SourceMap

Returns the CodeMap

pub fn span_err_with_help_and_notes<S: Into<MultiSpan> + Clone>(
&self,
sp: S,
msg: &str,
help: &Option<String>,
notes: &[(String, Option<S>)]
)

Emits an error message.

pub fn span_warn_with_help_and_notes<S: Into<MultiSpan> + Clone>(
&self,
sp: S,
msg: &str,
help: &Option<String>,
notes: &[(String, Option<S>)]
)

Emits an error message.

pub fn has_errors(&self) -> bool

Returns true if an error has been emitted

pub fn get_annotated_procedures(&self) -> Vec<ProcedureDefId>

Get ids of Rust procedures that are annotated with a Prusti specification

pub fn has_prusti_attribute(&self, def_id: ProcedureDefId, name: &str) -> bool

Find whether the procedure has a particular prusti::<name> attribute.

pub fn dump_borrowck_info(&self, procedures: &[ProcedureDefId])

Dump various information from the borrow checker.

Mostly used for experiments and debugging.

pub fn get_item_def_path(&self, def_id: DefId) -> String

Get an absolute def_path. Note: not preserved across compilations!

pub fn get_def_span(&self, def_id: DefId) -> Span

Get the span of the given definition.

pub fn get_absolute_item_name(&self, def_id: DefId) -> String

pub fn get_item_name(&self, def_id: DefId) -> String

pub fn get_procedure(&self, proc_def_id: ProcedureDefId) -> Procedure<'tcx>

Get a Procedure.

pub fn local_mir(&self, def_id: LocalDefId) -> Rc<Body<'tcx>>

Get the MIR body of a local procedure.

pub fn local_mir_borrowck_facts(&self, def_id: LocalDefId) -> Rc<BorrowckFacts>

Get Polonius facts of a local procedure.

pub fn external_mir<'a>(&self, def_id: DefId) -> &'a Body<'tcx>

Get the MIR body of an external procedure.

pub fn get_traits_decls_for_type(&self, ty: &Ty<'tcx>) -> HashSet<DefId>

Get all relevant trait declarations for some type.

pub fn get_assoc_item(&self, id: DefId, name: Symbol) -> Option<AssocItem>

Get an associated item by name.

pub fn get_trait_method_decl_for_type(
&self,
typ: Ty<'tcx>,
trait_id: DefId,
name: Symbol
) -> Vec<AssocItem>

Get a trait method declaration by name for type.

pub fn type_is_allowed_in_pure_functions(
&self,
ty: Ty<'tcx>,
param_env: ParamEnv<'tcx>
) -> bool

pub fn type_is_copy(&self, ty: Ty<'tcx>, param_env: ParamEnv<'tcx>) -> bool

pub fn type_implements_trait(
&self,
ty: Ty<'tcx>,
trait_def_id: DefId,
param_env: ParamEnv<'tcx>
) -> bool

Checks whether the given type implements the trait with the given DefId.