Struct prusti_interface::environment::ProcedureLoops

pub struct ProcedureLoops {
    pub loop_heads: FxHashSet<BasicBlockIndex>,
    pub loop_head_depths: FxHashMap<BasicBlockIndex, usize>,
    pub loop_bodies: FxHashMap<BasicBlockIndex, FxHashSet<BasicBlockIndex>>,
    pub ordered_loop_bodies: FxHashMap<BasicBlockIndex, Vec<BasicBlockIndex>>,
    pub back_edges: FxHashSet<(BasicBlockIndex, BasicBlockIndex)>,
    pub ordered_blocks: Vec<BasicBlockIndex>,
    /* private fields */
}

Struct that contains information about all loops in the procedure.

Fields
loop_heads: FxHashSet<BasicBlockIndex>

A list of basic blocks that are loop heads.

loop_head_depths: FxHashMap<BasicBlockIndex, usize>

The depth of each loop head, starting from one for a simple single loop.

loop_bodies: FxHashMap<BasicBlockIndex, FxHashSet<BasicBlockIndex>>

A map from loop heads to the corresponding bodies.

ordered_loop_bodies: FxHashMap<BasicBlockIndex, Vec<BasicBlockIndex>>back_edges: FxHashSet<(BasicBlockIndex, BasicBlockIndex)>

Back edges.

ordered_blocks: Vec<BasicBlockIndex>

The list of basic blocks ordered in the topological order (ignoring back edges).

Implementations

impl ProcedureLoops

pub fn new<'a, 'tcx: 'a>(
mir: &'a Body<'tcx>,
real_edges: &RealEdges
) -> ProcedureLoops

pub fn count_loop_heads(&self) -> usize

pub fn max_loop_nesting(&self) -> usize

pub fn is_loop_head(&self, bbi: BasicBlockIndex) -> bool

pub fn get_loop_exit_blocks(&self, bbi: BasicBlockIndex) -> &[BasicBlockIndex]

pub fn is_conditional_branch(
&self,
loop_head: BasicBlockIndex,
bbi: BasicBlockIndex
) -> bool

pub fn get_enclosing_loop_heads(
&self,
bbi: BasicBlockIndex
) -> &[BasicBlockIndex]

pub fn get_loop_head(&self, bbi: BasicBlockIndex) -> Option<BasicBlockIndex>

Get the loop head, if any Note: a loop head is loop head of itself

pub fn get_loop_head_depth(&self, bbi: BasicBlockIndex) -> usize

Get the depth of a loop head, starting from one for a simple loop

pub fn get_loop_depth(&self, bbi: BasicBlockIndex) -> usize

Get the loop-depth of a block (zero if it’s not in a loop).

pub fn get_loop_body(&self, loop_head: BasicBlockIndex) -> &[BasicBlockIndex]

Get the (topologically ordered) body of a loop, given a loop head

pub fn is_out_edge(&self, from: BasicBlockIndex, to: BasicBlockIndex) -> bool

Does this edge exit a loop?

pub fn is_block_in_loop(
&self,
loop_head: BasicBlockIndex,
block: BasicBlockIndex
) -> bool

Check if block is inside a given loop.

pub fn compute_read_and_write_leaves<'a, 'tcx: 'a>(
&self,
loop_head: BasicBlockIndex,
mir: &'a Body<'tcx>,
definitely_initalised_paths: Option<&PlaceSet<'tcx>>
) -> Result<(Vec<Place<'tcx>>, Vec<Place<'tcx>>, Vec<Place<'tcx>>), LoopAnalysisError>

If definitely_initalised_paths is not None, returns only leaves that are definitely initialised.