Module prusti_interface::specs

Modules

checker

external

typed

Structs

SpecCollector

Specification collector, intended to be applied as a visitor over the crate HIR. After the visit, SpecCollector::build_def_specs can be used to get back a mapping of DefIds (which may not be local due to extern specs) to their typed::SpecificationSet, i.e. procedures, loop invariants, and structs.