Struct prusti_interface::specs::SpecCollector

pub struct SpecCollector<'a, 'tcx> { /* private fields */ }

Specification collector, intended to be applied as a visitor over the crate HIR. After the visit, SpecCollector::build_def_specs can be used to get back a mapping of DefIds (which may not be local due to extern specs) to their [typed::SpecificationSet], i.e. procedures, loop invariants, and structs.

Implementations

impl<'a, 'tcx> SpecCollector<'a, 'tcx>

pub fn new(env: &'a mut Environment<'tcx>) -> Self

pub fn collect_specs(&mut self, hir: Map<'tcx>)

pub fn build_def_specs(&mut self) -> DefSpecificationMap