Module prusti_interface::specs::typed

Structs

DefSpecificationMap

A map of specifications keyed by crate-local DefIds.

GhostBegin

GhostEnd

Pledge

ProcedureSpecification

PrustiAssertion

PrustiAssumption

SpecGraph

The base container to store a contract of a procedure. A contract can be divided into multiple specifications:

SpecificationId

A unique ID of the specification element such as entire precondition or postcondition.

TypeSpecification

Specification of a type.

Enums

LoopSpecification

ProcedureSpecificationKind

ProcedureSpecificationKindError

SpecConstraintKind

SpecIdRef

A reference to a procedure specification.

SpecType

A specification type.

SpecificationItem

A specification, such as preconditions or a #[pure] annotation. Contains information about the refinement of these specifications.

Traits

Refinable