Trait prusti_interface::specs::typed::Spanned

pub trait Spanned<'tcx> {
    fn get_spans(&self, mir_body: &Body<'tcx>, tcx: TyCtxt<'tcx>) -> Vec<Span>;
}

This trait is implemented for specification-related types that have one or more associated spans (positions within the source code). The spans are not necessarily contiguous, and may be used for diagnostic reporting.

Required methods

fn get_spans(&self, mir_body: &Body<'tcx>, tcx: TyCtxt<'tcx>) -> Vec<Span>

Returns the spans for the given value. mir is the function body used to resolve positions of rustc_middle::mir::Local indices, tcx is used to resolve positions of global items.