Function prusti_interface::utils::expand

pub fn expand<'tcx>(
    mir: &Body<'tcx>,
    tcx: TyCtxt<'tcx>,
    minuend: Place<'tcx>,
    subtrahend: Place<'tcx>
) -> Vec<Place<'tcx>>

Subtract the subtrahend place from the minuend place. The subtraction is defined as set minus between minuend place replaced with a set of places that are unrolled up to the same level as subtrahend and the singleton subtrahend set. For example, subtract(x.f, x.f.g.h) is performed by unrolling x.f into {x.g, x.h, x.f.f, x.f.h, x.f.g.f, x.f.g.g, x.f.g.h} and subtracting {x.f.g.h} from it, which results into {x.g, x.h, x.f.f, x.f.h, x.f.g.f, x.f.g.g}.