Function prusti_interface::environment::mir_body::borrowck::facts::patch::apply_patch_to_borrowck

pub fn apply_patch_to_borrowck<'tcx>(
    borrowck_input_facts: &mut AllInputFacts,
    location_table: &mut LocationTable,
    patch: &MirPatch<'tcx>,
    old_body: &Body<'tcx>,
    patched_body: &mut Body<'tcx>
)

FIXME: Currently, this function patches only borrowck_input_facts.cfg_edge. It probably should also patch other facts. For example, added drop flag assignment statements may mark some lifetimes as dead and result in premature ending of borrows. Potential example:

fn test6() {
    let a = T { f: 4 };
    let b = T3 { g: a };
    let mut c = 4;
    let x = &mut c;
    if random() {
        drop(b.g);
        *x = 5;
    }
}