Module prusti_interface::environment::polonius_info

Structs

AdditionalFacts

Additional facts derived from the borrow checker facts.

LoanPlaces

This represents the assignment in which a loan was created. The source will contain the creation of the loan, while the dest will store the created reference.

LoopMagicWand

We are guaranteed to have only the permissions that are currently borrowed by the variable. For instance, in the example below x borrows one variable while the other one is consumed. As a result, we can restore the permissions only to the variable which we know was borrowed by x.

PoloniusInfo

ReborrowingDAG

ReborrowingDAGNode

Enums

PoloniusInfoError

ReborrowingKind

ReborrowingZombity

Functions

graphviz