Struct prusti_interface::environment::polonius_info::PoloniusInfo

pub struct PoloniusInfo<'a, 'tcx: 'a> {
    pub place_regions: PlaceRegions,
    // some fields omitted
}

Fields
place_regions: PlaceRegions
Implementations

impl<'a, 'tcx: 'a> PoloniusInfo<'a, 'tcx>

pub fn new(
env: &'a Environment<'tcx>,
procedure: &'a Procedure<'tcx>,
_loop_invariant_block: &FxHashMap<BasicBlock, BasicBlock>
) -> Result<Self, PoloniusInfoError>

pub fn get_point(&self, location: Location, point_type: PointType) -> PointIndex

pub fn get_all_loans_kept_alive_by(
&self,
point: PointIndex,
region: Region
) -> (Vec<Loan>, Vec<Loan>)

pub fn get_all_active_loans(&self, location: Location) -> (Vec<Loan>, Vec<Loan>)

Get loans including the zombies (all_loans, zombie_loans).

pub fn get_active_loans(&self, location: Location, zombie: bool) -> Vec<Loan>

Get loans that are active (including those that are about to die) at the given location.

pub fn get_all_loans_dying_at(
&self,
location: Location
) -> (Vec<Loan>, Vec<Loan>)

Get loans including the zombies (all_loans, zombie_loans).

pub fn get_all_loans_dying_between(
&self,
initial_loc: Location,
final_loc: Location
) -> (Vec<Loan>, Vec<Loan>)

Get loans including the zombies (all_loans, zombie_loans).

pub fn get_loans_dying_at(&self, location: Location, zombie: bool) -> Vec<Loan>

Get loans that die at (that is, exactly after) the given location.

pub fn get_loans_dying_between(
&self,
initial_loc: Location,
final_loc: Location,
zombie: bool
) -> Vec<Loan>

Get loans that die between two consecutive locations

pub fn get_conflicting_loans(&self, loan: Loan) -> Vec<Loan>

pub fn get_alive_conflicting_loans(
&self,
loan: Loan,
location: Location
) -> Vec<Loan>

pub fn get_loan_location(&self, loan: &Loan) -> Location

pub fn get_loan_at_location(&self, location: Location) -> Loan

pub fn get_call_loan_at_location(&self, location: Location) -> Option<Loan>

pub fn loan_locations(&self) -> FxHashMap<Loan, Location>

pub fn get_loan_places(
&self,
loan: &Loan
) -> Result<Option<LoanPlaces<'tcx>>, PlaceRegionsError>

Convert a facts::Loan to LoanPlaces<’tcx> (if possible)

pub fn get_assignment_for_loan(
&self,
loan: Loan
) -> Result<Option<Statement<'tcx>>, PlaceRegionsError>

Returns the atomic assignment that created a loan. Refer to the documentation of SplitAggregateAssignment for more information on what an atomic assignment is.

pub fn find_variable(&self, _region: Region) -> Option<Local>

Find a variable that has the given region in its type.

pub fn construct_reborrowing_dag(
&self,
loans: &[Loan],
zombie_loans: &[Loan],
location: Location
) -> Result<ReborrowingDAG, PoloniusInfoError>

loans – all loans, including the zombie loans.

pub fn construct_reborrowing_dag_loop_body(
&self,
loans: &[Loan],
zombie_loans: &[Loan],
location: Location
) -> Result<ReborrowingDAG, PoloniusInfoError>

pub fn get_loan_loops(
&self,
loans: &[Loan]
) -> Result<Vec<(Loan, BasicBlock)>, PoloniusInfoError>

Get loops in which loans are defined (if any).

pub fn construct_reborrowing_dag_custom_reborrows(
&self,
loans: &[Loan],
zombie_loans: &[Loan],
location: Location,
reborrows_direct: &[(Loan, Loan)]
) -> Result<ReborrowingDAG, PoloniusInfoError>

loans – all loans, including the zombie loans.