Struct prusti_interface::environment::Environment

pub struct Environment<'tcx> {
    pub body: EnvBody<'tcx>,
    pub diagnostic: EnvDiagnostic<'tcx>,
    pub name: EnvName<'tcx>,
    pub query: EnvQuery<'tcx>,
    /* private fields */
}

Facade to the Rust compiler.

Fields
body: EnvBody<'tcx>diagnostic: EnvDiagnostic<'tcx>name: EnvName<'tcx>query: EnvQuery<'tcx>
Implementations

impl<'tcx> Environment<'tcx>

pub fn new(tcx: TyCtxt<'tcx>, prusti_version: &'static str) -> Self

Builds an environment given a compiler state.

pub fn tcx(&self) -> TyCtxt<'tcx>

Returns the typing context

pub fn dump_borrowck_info(&self, procedures: &[ProcedureDefId])

Dump various information from the borrow checker.

Mostly used for experiments and debugging.

pub fn get_procedure(&self, proc_def_id: ProcedureDefId) -> Procedure<'tcx>

Get a Procedure.

pub fn get_annotated_procedures_and_types(
&self
) -> (Vec<ProcedureDefId>, Vec<Ty<'tcx>>)

Get ids of Rust procedures that are annotated with a Prusti specification

pub fn compare_prusti_version(&self, other: &str) -> Ordering

Compare the current version of the prusti crate to the given other version

pub fn get_prusti_version(&self) -> &'static str

Get the current version of the prusti crate

pub fn compare_to_curr_specs_version(other: &str) -> Ordering

Compare to the current version of the prusti-specs crate which we’ve been compiled with

pub fn get_specs_version() -> &'static str

Get the current version of the prusti crate