Struct prusti_interface::environment::Environment

pub struct Environment<'tcx> { /* private fields */ }

Facade to the Rust compiler.

Implementations

impl<'tcx> Environment<'tcx>

pub fn new(tcx: TyCtxt<'tcx>) -> Self

Builds an environment given a compiler state.

pub fn source_path(&self) -> PathBuf

Returns the path of the source that is being compiled

pub fn source_file_name(&self) -> String

Returns the file name of the source that is being compiled

pub fn crate_name(&self) -> String

Returns the name of the crate that is being compiled

pub fn tcx(&self) -> TyCtxt<'tcx>

Returns the typing context

pub fn hir_id_to_type(&self, hir_id: HirId) -> Ty<'tcx>

Returns the type of a HirId

pub fn codemap(&self) -> &'tcx SourceMap

Returns the CodeMap

pub fn span_err_with_help_and_notes<S: Into<MultiSpan> + Clone>(
&self,
sp: S,
msg: &str,
help: &Option<String>,
notes: &[(String, Option<S>)]
)

Emits an error message.

pub fn span_warn_with_help_and_notes<S: Into<MultiSpan> + Clone>(
&self,
sp: S,
msg: &str,
help: &Option<String>,
notes: &[(String, Option<S>)]
)

Emits a warning message.

pub fn span_warn_on_err_with_help_and_notes<S: Into<MultiSpan> + Clone>(
&self,
sp: S,
msg: &str,
help: &Option<String>,
notes: &[(String, Option<S>)]
)

Buffers a warning message, to be emitted on error.

pub fn has_errors(&self) -> bool

Returns true if an error has been emitted

pub fn get_annotated_procedures(&self) -> Vec<ProcedureDefId>

Get ids of Rust procedures that are annotated with a Prusti specification

pub fn get_local_attributes(&self, def_id: LocalDefId) -> &[Attribute]

pub fn get_attributes(&self, def_id: ProcedureDefId) -> &[Attribute]

pub fn has_prusti_attribute(&self, def_id: ProcedureDefId, name: &str) -> bool

Find whether the procedure has a particular prusti::<name> attribute.

pub fn dump_borrowck_info(&self, procedures: &[ProcedureDefId])

Dump various information from the borrow checker.

Mostly used for experiments and debugging.

pub fn get_item_def_path(&self, def_id: DefId) -> String

Get an absolute def_path. Note: not preserved across compilations!

pub fn get_unique_item_name(&self, def_id: DefId) -> String

Get descriptive name prepended with crate name to make it unique.

pub fn get_def_span(&self, def_id: DefId) -> Span

Get the span of the given definition.

pub fn get_absolute_item_name(&self, def_id: DefId) -> String

pub fn get_item_name(&self, def_id: DefId) -> String

pub fn get_procedure(&self, proc_def_id: ProcedureDefId) -> Procedure<'tcx>

Get a Procedure.

pub fn local_mir(
&self,
def_id: LocalDefId,
substs: SubstsRef<'tcx>
) -> Rc<Body<'tcx>>

Get the MIR body of a local procedure, monomorphised with the given type substitutions.

pub fn local_mir_borrowck_facts(&self, def_id: LocalDefId) -> Rc<BorrowckFacts>

Get Polonius facts of a local procedure.

pub fn try_get_local_mir_borrowck_facts(
&self,
def_id: LocalDefId
) -> Option<Rc<BorrowckFacts>>

pub fn external_mir(
&self,
def_id: DefId,
substs: SubstsRef<'tcx>
) -> Rc<Body<'tcx>>

Get the MIR body of an external procedure.

pub fn get_traits_decls_for_type(&self, ty: &Ty<'tcx>) -> HashSet<DefId>

Get all relevant trait declarations for some type.

pub fn get_assoc_item(&self, id: DefId, name: Symbol) -> Option<AssocItem>

Get an associated item by name.

pub fn is_trait_method(&self, def_id: ProcedureDefId) -> bool

Returns true iff def_id is a trait method

pub fn is_trait_method_impl(&self, def_id: ProcedureDefId) -> bool

Returns true iff def_id is an implementation of a trait method

pub fn find_trait_method(&self, impl_def_id: ProcedureDefId) -> Option<DefId>

Returns the DefId of the corresponding trait method, if any. This should not be used to resolve calls (where substs are known): use find_trait_method_substs instead!

pub fn find_trait_method_substs(
&self,
impl_method_def_id: ProcedureDefId,
impl_method_substs: SubstsRef<'tcx>
) -> Option<(ProcedureDefId, SubstsRef<'tcx>)>

If the given impl_method_def_id is an implementation of a trait method, return the DefId of that trait method as well as an adapted version of the callsite impl_method_substs substitutions.

pub fn find_impl_of_trait_method_call(
&self,
proc_def_id: ProcedureDefId,
substs: SubstsRef<'tcx>
) -> Option<ProcedureDefId>

Given some procedure proc_def_id which is called, this method returns the actual method which will be executed when proc_def_id is defined on a trait. Returns None if this method can not be found or the provided proc_def_id is no trait item.

pub fn resolve_method_call(
&self,
caller_def_id: ProcedureDefId,
called_def_id: ProcedureDefId,
call_substs: SubstsRef<'tcx>
) -> (ProcedureDefId, SubstsRef<'tcx>)

Given a call to called_def_id from within caller_def_id, returns the DefId that will actually be called if known (i.e. if a trait method call actually resolves to a concrete implementation), as well as the correct substitutions for that call. If a method is not resolved, returns the original called_def_id and call_substs.

pub fn type_is_copy(
&self,
ty: Binder<'tcx, Ty<'tcx>>,
param_env: ParamEnv<'tcx>
) -> bool

Checks whether ty is copy. The type is wrapped into a Binder to handle regions correctly.

pub fn type_implements_trait(
&self,
ty: Ty<'tcx>,
trait_def_id: DefId,
param_env: ParamEnv<'tcx>
) -> bool

Checks whether the given type implements the trait with the given DefId.

pub fn type_implements_trait_with_trait_substs(
&self,
ty: Ty<'tcx>,
trait_def_id: DefId,
trait_substs: SubstsRef<'tcx>,
param_env: ParamEnv<'tcx>
) -> bool

Checks whether the given type implements the trait with the given DefId. Accounts for generic params on the trait by the given trait_substs.

pub fn identity_substs(&self, def_id: ProcedureDefId) -> SubstsRef<'tcx>

Return the default substitutions for a particular item, i.e. where each generic maps to itself.

pub fn evaluate_predicate(
&self,
predicate: Predicate<'tcx>,
param_env: ParamEnv<'tcx>
) -> bool

Evaluates the provided ty::Predicate. Returns true if the predicate is fulfilled. Returns false if the predicate is not fulfilled or it could not be evaluated.

pub fn resolve_assoc_types<T: TypeFoldable<'tcx> + Debug + Copy>(
&self,
normalizable: T,
param_env: ParamEnv<'tcx>
) -> T

Normalizes associated types in foldable types, i.e. this resolves projection types (ty::TyKind::Projections) Important: Regions while be erased during this process