Module prusti_interface::specs

Modules

checker

Module for verifying user-provided specifications after macro expansion

cross_crate

decoder

encoder

external

typed

Structs

SpecCollector

Specification collector, intended to be applied as a visitor over the crate HIR. After the visit, SpecCollector::build_def_specs can be used to get back a mapping of DefIds (which may not be local due to extern specs) to their [typed::SpecificationSet], i.e. procedures, loop invariants, and structs.

Functions

is_spec_fn

Returns true iff def_id points to a spec function (i.e. a function for which we don’t need polonius/borrowck facts)