Struct prusti_interface::specs::external::ExternSpecResolver

pub struct ExternSpecResolver<'tcx> {
    pub extern_fn_map: HashMap<ExternSpecDeclaration, DefId>,
    /* private fields */
}

This struct is used to build a mapping of external functions to their Prusti specifications (see extern_fn_map).

Fields
extern_fn_map: HashMap<ExternSpecDeclaration, DefId>

Maps real functions to Prusti-generated fake functions with specifications.

Implementations

impl<'tcx> ExternSpecResolver<'tcx>

pub fn new(env: &Environment<'tcx>) -> Self

pub fn add_extern_fn(
&mut self,
fn_kind: FnKind<'tcx>,
fn_decl: &'tcx FnDecl<'_>,
body_id: BodyId,
span: Span,
id: HirId,
extern_spec_kind: ExternSpecKind
)

Registers an external function specification. The arguments for this function are the same as arguments given to a function visit in an intravisit visitor.

In case of duplicates, the function is added to spec_duplicates, and will later (in check_duplicates) be reported as an error. Otherwise, the function is added to extern_fn_map.

pub fn check_errors(&self, env_diagnostic: &EnvDiagnostic<'tcx>)

Report errors encountered when resolving extern specs