Module prusti_interface::specs::typed

Structs

DefSpecificationMap

A map of specifications keyed by crate-local DefIds.

ExpressionId

A unique ID of the Rust expression used in the specification.

SpecificationId

A unique ID of the specification element such as entire precondition or postcondition.

Enums

SpecIdRef

A reference to a procedure specification.

SpecType

A specification type.

Traits

Spanned

This trait is implemented for specification-related types that have one or more associated spans (positions within the source code). The spans are not necessarily contiguous, and may be used for diagnostic reporting.

StructuralToTyped

Type Definitions

Assertion

An assertion that has no types associated with it.

AssertionKind

An assertion kind that has no types associated with it.

Expression

An expression that has no types associated with it.

LoopSpecification

A set of untyped specifications associated with a loop.

Pledge

A pledge in the postcondition.

ProcedureSpecification

A set of untyped specifications associated with a procedure.

QuantifierVars

Quantifier variables that have no types associated with it.

SpecEntailmentVars

Specification entailment variables that have no types associated.

Specification

A specification that has no types associated with it.

SpecificationMap

A map of untyped specifications for a specific crate.

SpecificationSet

A set of untyped specifications associated with a single element.

Trigger

A trigger that has no types associated with it.

TriggerSet

A trigger set that has no types associated with it.